hpcviewer works, hpctraceviewer broken

Conflicts:
	edu.rice.cs.hpc.data/src/edu/rice/cs/hpc/data/experiment/scope/visitors/FlatViewScopeVisitor.java
	edu.rice.cs.hpc.viewer/META-INF/MANIFEST.MF
	edu.rice.cs.hpc.viewer/hpcviewer.plugin.product
	edu.rice.cs.hpc.viewer/plugin.xml
	edu.rice.cs.hpc.viewer/src/edu/rice/cs/hpc/viewer/experiment/ExperimentManager.java

+++ edu.rice.cs.hpc.data/src/edu/rice/cs/hpc/data/db/DataSummary.java
@@ -227,7 +227,7 @@ public class DataSummary extends DataCommon
 		int offset = 0;
 		for (int i=0; i<num_cctid; i++)
 		{
-			int offset_size  = (int) (cct_table[i+1] - cct_table[i]);
+			int offset_size  = (cct_table[i+1] - cct_table[i]);
 			int num_metrics  = offset_size / METRIC_ENTRY_SIZE;
 			
 			metric_id[i] 	 = new int  [num_metrics];
